Add training records

+++ from github/Stock-Trading-Environment/env/StockTradingEnv.py
@@ -84,8 +84,8 @@ class StockTradingEnv(gym.Env):
             # Buy amount % of balance in shares
             
 
-            total_possible = int(self.balance / self.actual_price)
-            shares_bought = int(total_possible * amount)
+            total_possible = self.balance / self.actual_price
+            shares_bought = total_possible * amount
             self.current_action = shares_bought
             prev_cost = self.cost_basis * self.shares_held
             additional_cost = shares_bought * self.actual_price * (1+COMMISSION_FEE)
@@ -98,7 +98,7 @@ class StockTradingEnv(gym.Env):
         elif action_type < 2:
             # Sell amount % of shares held
 
-            shares_sold = int(self.shares_held * amount)
+            shares_sold = self.shares_held * amount
             self.current_action = shares_sold*-1
             self.balance += shares_sold * self.actual_price
             self.shares_held -= shares_sold
